Pushing tuesday's changes to the scirpts
This is principally a formatting change, especially to T2.py. This has cleaned up a lot of the code, adding titles to graphs, making it clearner to read, so all good stuff.

Also the pulse height as a function of period was moved out of PNMR.py and into a seperate file so to make it cleaner to read :)